update week clean

+++ src/main/java/org/soraworld/ranklist/util/DateUtils.java
@@ -7,25 +7,25 @@ import java.util.Date;
  */
 public class DateUtils {
 
-    public static long getDay(Date date,int timezone) {
-        long timestamp = date.getTime() + 3600000*timezone;
+    public static long getDay(Date date, int timezone) {
+        long timestamp = date.getTime() + 3600000 * timezone;
         return timestamp / 86400000L + 1;
     }
 
-    public static long getDay(long timestamp,int timezone) {
-        return (timestamp + 3600000*timezone) / 86400000L + 1;
+    public static long getDay(long timestamp, int timezone) {
+        return (timestamp + 3600000 * timezone) / 86400000L + 1;
     }
 
-    public static long getWeek(Date date,int timezone) {
-        long days = getDay(date,timezone) + 3;
-        return (days + 6) / 7;
+    public static int getWeek(Date date, int timezone) {
+        long days = getDay(date, timezone) + 3;
+        return (int) ((days + 6) / 7);
     }
 
-    public static int getWeekDay(Date date,int timezone) {
-        long days = getDay(date,timezone) + 3;
+    public static int getWeekDay(Date date, int timezone) {
+        long days = getDay(date, timezone) + 3;
         return (int) (days + 6) % 7 + 1;
     }
-    
+
     public static long getDay(Date date) {
         long timestamp = date.getTime() + 28800000L;
         return timestamp / 86400000L + 1;
@@ -35,9 +35,9 @@ public class DateUtils {
         return (timestamp + 28800000L) / 86400000L + 1;
     }
 
-    public static long getWeek(Date date) {
+    public static int getWeek(Date date) {
         long days = getDay(date) + 3;
-        return (days + 6) / 7;
+        return (int) ((days + 6) / 7);
     }
 
     public static int getWeekDay(Date date) {
